All arithmetic tests passing

+++ src/main/java/com/zipcodewilmington/scientificcalculator/Calculator.java
@@ -1,9 +1,7 @@
 package com.zipcodewilmington.scientificcalculator;
 
-import java.sql.SQLOutput;
-import java.util.Scanner;
-
-import static com.zipcodewilmington.scientificcalculator.MainApplication.isNumeric;
+//import java.sql.SQLOutput;
+//import static com.zipcodewilmington.scientificcalculator.MainApplication.isNumeric;
 
 
 import java.util.Scanner;
@@ -24,38 +22,38 @@ public class Calculator {
         if (userInput.equals("+")) {
             int y = getYInput();
             int x = getXInput();
-            double sum = add(y, x);
+            int sum = add(y, x);
             System.out.println(sum);
         } else if (userInput.equals("-")) {
             int y = getYInput();
             int x = getXInput();
-            double difference = subtract(y, x);
+            int difference = subtract(y, x);
             System.out.println(difference);
         } else if (userInput.equals("*")) {
             int y = getYInput();
             int x = getXInput();
-            double product = multiply(y, x);
+            int product = multiply(y, x);
             System.out.println(product);
         } else if (userInput.equals("/")) {
             int y = getYInput();
             int x = getXInput();
-            double quotient = divide(y, x);
+            int quotient = divide(y, x);
             System.out.println(quotient);
         }
     }
 
-    public double add(int int1, int int2) {
+    public int add(int int1, int int2) {
         return (int1 + int2);
     }
 
-    public double subtract(int int1, int int2) {
+    public int subtract(int int1, int int2) {
         return (int1 - int2);
     }
 
-    public double multiply(int int1, int int2) {
+    public int multiply(int int1, int int2) {
         return (int1 * int2);
     }
-    public double divide(int int1, int int2) {
+    public int divide(int int1, int int2) {
         return (int1 / int2);
     }
 
@@ -78,25 +76,3 @@ public class Calculator {
         System.out.println(lastInput);
     }
 }
-
-
-
-
-
-//        String yValueInput = "";
-//        double yValue = 0;
-//        double result = 0;
-//        if (userInput.equals("+")) {
-//            System.out.println("Enter second number: ");
-//            yValueInput = scanner.nextLine();
-//            if (isNumeric(yValueInput)) {
-//                yValue = Double.parseDouble(yValueInput);
-//                result = display + yValue;
-//                return result; //result of x + y and returns it calc.run();
-//            }
-//        }
-//        return display;//checks if user inputs out of bounds input then method spits out original number input from beginning
-
-
-
-
